LayoutTests/imported/w3c/web-platform-tests/webrtc/toJSON-expected.txt


PASS RTCSessionDescription.toJSON serializes only specific attributes
FAIL RTCIceCandidate.toJSON serializes only specific attributes assert_equals: expected (string) "test" but got (object) null
